refactor with errors
