+++ admin/routes/school.js
@@ -5,10 +5,6 @@ const { getHome,
   getPupils,
   getResults,
   downloadResults,
-  getGeneratePinsOverview,
-  getGeneratePinsList,
-  postGeneratePins,
-  getGeneratedPinsList,
   getSubmitAttendance,
   postSubmitAttendance,
   getDeclarationForm,
@@ -24,10 +20,6 @@ const school = (router) => {
   router.get('/pupil-register/:sortColumn/:sortOrder', isAuthenticated(config.ROLE_TEACHER), (req, res, next) => getPupils(req, res, next))
   router.get('/results', isAuthenticated(config.ROLE_TEACHER), (req, res, next) => getResults(req, res, next))
   router.get('/download-results', isAuthenticated(config.ROLE_TEACHER), (req, res, next) => downloadResults(req, res, next))
-  router.get('/generate-pins-overview', isAuthenticated(config.ROLE_TEACHER), (req, res, next) => getGeneratePinsOverview(req, res, next))
-  router.get('/generate-pins-list/:sortField/:sortDirection', isAuthenticated(config.ROLE_TEACHER), (req, res, next) => getGeneratePinsList(req, res, next))
-  router.post('/generate-pins', isAuthenticated(config.ROLE_TEACHER), (req, res, next) => postGeneratePins(req, res, next))
-  router.get('/generated-pins-list', isAuthenticated(config.ROLE_TEACHER), (req, res, next) => getGeneratedPinsList(req, res, next))
   router.get('/submit-attendance', isAuthenticated(config.ROLE_TEACHER), (req, res, next) => getSubmitAttendance(req, res, next))
   router.post('/submit-attendance-form', isAuthenticated(config.ROLE_TEACHER), (req, res, next) => postSubmitAttendance(req, res, next))
   router.get('/declaration-form', isAuthenticated(config.ROLE_TEACHER), (req, res, next) => getDeclarationForm(req, res, next))
