+++ admin/spec/controllers/pupil-pin.spec.js
@@ -1,0 +1,282 @@
+'use strict'
+/* global describe beforeEach afterEach it expect jasmine spyOn */
+const proxyquire = require('proxyquire')
+const httpMocks = require('node-mocks-http')
+const sinon = require('sinon')
+require('sinon-mongoose')
+
+const config = require('../../config')
+const School = require('../../models/school')
+const schoolDataService = require('../../services/data-access/school.data.service')
+const pupilDataService = require('../../services/data-access/pupil.data.service')
+const schoolService = require('../../services/school.service')
+const pupilService = require('../../services/pupil.service')
+const generatePinsService = require('../../services/generate-pins.service')
+const sortingAttributesService = require('../../services/sorting-attributes.service')
+const dateService = require('../../services/date.service')
+const qrService = require('../../services/qr.service')
+
+describe('pupilPin controller:', () => {
+  function getRes () {
+    const res = httpMocks.createResponse()
+    res.locals = {}
+    return res
+  }
+
+  function getReq (params) {
+    const req = httpMocks.createRequest(params)
+    req.user = { School: 9991001 }
+    req.breadcrumbs = jasmine.createSpy('breadcrumbs')
+    req.flash = jasmine.createSpy('flash')
+    return req
+  }
+
+  describe('getGeneratePinsOverview route', () => {
+    let sandbox
+    let goodReqParams = {
+      method: 'GET',
+      url: '/pupil-pin/generate-pins-overview',
+      session: {
+        id: 'ArRFdOiz1xI8w0ljtvVuD6LU39pcfgqy'
+      }
+    }
+
+    beforeEach(() => {
+      sandbox = sinon.sandbox.create()
+    })
+
+    afterEach(() => {
+      sandbox.restore()
+    })
+
+    it('displays the generate pins overview page if no active pins are present', async (done) => {
+      const res = getRes()
+      const req = getReq(goodReqParams)
+      const controller = require('../../controllers/pupil-pin').getGeneratePinsOverview
+      spyOn(res, 'render').and.returnValue(null)
+      spyOn(pupilService, 'getPupilsWithActivePins').and.returnValue([])
+      await controller(req, res)
+      expect(res.locals.pageTitle).toBe('Generate pupil PINs')
+      expect(res.render).toHaveBeenCalled()
+      done()
+    })
+    it('displays the generated pins list page if active pins are present', async (done) => {
+      const res = getRes()
+      const req = getReq(goodReqParams)
+      const controller = require('../../controllers/pupil-pin').getGeneratePinsOverview
+      spyOn(res, 'redirect').and.returnValue(null)
+      spyOn(pupilService, 'getPupilsWithActivePins').and.returnValue([ '1', '2' ])
+      await controller(req, res)
+      expect(res.redirect).toHaveBeenCalled()
+      done()
+    })
+  })
+
+  describe('getGeneratePinsList route', () => {
+    let sandbox
+    let next
+    let controller
+    let goodReqParams = {
+      method: 'GET',
+      url: '/pupil-pin/generate-pins-list',
+      session: {
+        id: 'ArRFdOiz1xI8w0ljtvVuD6LU39pcfgqy'
+      }
+    }
+
+    beforeEach(() => {
+      sandbox = sinon.sandbox.create()
+      next = jasmine.createSpy('next')
+    })
+
+    afterEach(() => {
+      sandbox.restore()
+    })
+
+    describe('when the school is found in the database', () => {
+      beforeEach(() => {
+        sandbox.mock(schoolDataService).expects('findOne').resolves(new School({ name: 'Test School' }))
+        controller = proxyquire('../../controllers/pupil-pin.js', {
+          '../../services/data-access/school.data.service': schoolDataService
+        }).getGeneratePinsList
+      })
+
+      it('displays the generate pins list page', async (done) => {
+        const res = getRes()
+        const req = getReq(goodReqParams)
+        spyOn(generatePinsService, 'getPupils').and.returnValue(Promise.resolve({}))
+        spyOn(sortingAttributesService, 'getAttributes')
+          .and.returnValue({
+            htmlSortDirection: { lastName: 'asc' },
+            arrowSortDirection: { lastName: 'sort' }
+          })
+        spyOn(res, 'render').and.returnValue(null)
+        await controller(req, res, next)
+        expect(res.locals.pageTitle).toBe('Select pupils')
+        expect(res.render).toHaveBeenCalled()
+        done()
+      })
+    })
+
+    describe('when the school is not found in the database', () => {
+      beforeEach(() => {
+        sandbox.mock(schoolDataService).expects('findOne').resolves(null)
+        controller = proxyquire('../../controllers/pupil-pin.js', {
+          '../../services/data-access/school.data.service': schoolDataService
+        }).getGeneratePinsList
+      })
+      it('it throws an error', async (done) => {
+        const res = getRes()
+        const req = getReq(goodReqParams)
+        await controller(req, res, next)
+        expect(next).toHaveBeenCalled()
+        done()
+      })
+    })
+  })
+
+  describe('postGeneratePins route', () => {
+    let sandbox
+    let next
+    let goodReqParams = {
+      method: 'POST',
+      url: '/pupil-pin/generate-pins-list',
+      session: {
+        id: 'ArRFdOiz1xI8w0ljtvVuD6LU39pcfgqy'
+      },
+      body: {
+        pupil: [ '595cd5416e5ca13e48ed2519', '595cd5416e5ca13e48ed2520' ]
+      }
+    }
+
+    beforeEach(() => {
+      sandbox = sinon.sandbox.create()
+      next = jasmine.createSpy('next')
+    })
+
+    afterEach(() => {
+      sandbox.restore()
+    })
+    it('displays the generated pins list page after successful saving', async (done) => {
+      const res = getRes()
+      const req = getReq(goodReqParams)
+      const controller = require('../../controllers/pupil-pin.js').postGeneratePins
+      spyOn(generatePinsService, 'generatePupilPins').and.returnValue(null)
+      spyOn(pupilDataService, 'updateMultiple').and.returnValue(true)
+      spyOn(schoolDataService, 'findOne').and.returnValue(new School({ _id: 1, name: 'Test School' }))
+      spyOn(generatePinsService, 'generateSchoolPassword').and.returnValue({ schoolPin: '', pinExpiresAt: '' })
+      spyOn(schoolDataService, 'update').and.returnValue(null)
+      spyOn(res, 'redirect').and.returnValue(null)
+      await controller(req, res, next)
+      expect(res.redirect).toHaveBeenCalledWith('/pupil-pin/generated-pins-list')
+      done()
+    })
+    it('displays the generate pins list page if no pupil list is provided', async (done) => {
+      const res = getRes()
+      const req = { body: {} }
+      const controller = require('../../controllers/pupil-pin.js').postGeneratePins
+      spyOn(generatePinsService, 'generatePupilPins').and.returnValue(null)
+      spyOn(pupilDataService, 'updateMultiple').and.returnValue(true)
+      spyOn(generatePinsService, 'generateSchoolPassword').and.returnValue(null)
+      spyOn(res, 'redirect').and.returnValue(null)
+      await controller(req, res, next)
+      expect(res.redirect).toHaveBeenCalledWith('/pupil-pin/generate-pins-list')
+      done()
+    })
+    it('calls next with an error if school is not found', async (done) => {
+      const res = getRes()
+      const req = getReq(goodReqParams)
+      const controller = require('../../controllers/pupil-pin.js').postGeneratePins
+      spyOn(generatePinsService, 'generatePupilPins').and.returnValue(null)
+      spyOn(pupilDataService, 'updateMultiple').and.returnValue(true)
+      spyOn(schoolDataService, 'findOne').and.returnValue(null)
+      await controller(req, res, next)
+      expect(next).toHaveBeenCalled()
+      done()
+    })
+  })
+
+  describe('getGeneratedPins route', () => {
+    let sandbox
+    let next
+    let goodReqParams = {
+      method: 'POST',
+      url: '/pupil-pin/generate-pins-list',
+      session: {
+        id: 'ArRFdOiz1xI8w0ljtvVuD6LU39pcfgqy'
+      }
+    }
+
+    beforeEach(() => {
+      sandbox = sinon.sandbox.create()
+      next = jasmine.createSpy('next')
+    })
+
+    afterEach(() => {
+      sandbox.restore()
+    })
+    it('displays the generated pupils list and password', async (done) => {
+      const res = getRes()
+      const req = getReq(goodReqParams)
+      const controller = require('../../controllers/pupil-pin.js').getGeneratedPinsList
+      spyOn(pupilService, 'getPupilsWithActivePins').and.returnValue(null)
+      spyOn(schoolService, 'getActiveSchool').and.returnValue(null)
+      spyOn(res, 'render').and.returnValue(null)
+      await controller(req, res, next)
+      expect(res.render).toHaveBeenCalled()
+      done()
+    })
+    it('calls next if error occurs', async (done) => {
+      const res = getRes()
+      const req = getReq(goodReqParams)
+      const controller = require('../../controllers/pupil-pin.js').getGeneratedPinsList
+      spyOn(pupilService, 'getPupilsWithActivePins').and.returnValue(Promise.reject(new Error('error')))
+      await controller(req, res, next)
+      expect(next).toHaveBeenCalled()
+      done()
+    })
+  })
+
+  describe('getPrintPins route', () => {
+    let controller
+    let sandbox
+    let next
+    let goodReqParams = {
+      method: 'GET',
+      url: '/pupil/download-error-csv',
+      session: {
+        id: 'ArRFdOiz1xI8w0ljtvVuD6LU39pcfgqy'
+      }
+    }
+
+    beforeEach(() => {
+      sandbox = sinon.sandbox.create()
+      next = jasmine.createSpy('next')
+      controller = require('../../controllers/pupil-pin.js').getPrintPins
+    })
+
+    afterEach(() => {
+      sandbox.restore()
+    })
+
+    it('returns data for the print page', async (done) => {
+      spyOn(dateService, 'formatDayAndDate').and.returnValue('')
+      spyOn(pupilService, 'getPupilsWithActivePins').and.returnValue([])
+      spyOn(schoolService, 'getActiveSchool').and.returnValue({})
+      spyOn(qrService, 'getDataURL').and.returnValue('')
+      const res = getRes()
+      const req = getReq(goodReqParams)
+      spyOn(res, 'render').and.returnValue(null)
+      await controller(req, res, next)
+      expect(res.statusCode).toBe(200)
+      expect(res.render).toHaveBeenCalledWith('pupil-pin/pin-print', {
+        pupils: [],
+        school: {},
+        date: '',
+        qrDataURL: '',
+        url: config.PUPIL_APP_URL
+      })
+      done()
+    })
+  })
+})
