+++ admin/spec/service/generate-pins.service.spec.js
@@ -5,7 +5,6 @@ const proxyquire = require('proxyquire').noCallThru()
 const moment = require('moment')
 const sinon = require('sinon')
 const pupilDataService = require('../../services/data-access/pupil.data.service')
-const schoolDataService = require('../../services/data-access/school.data.service')
 const generatePinsService = require('../../services/generate-pins.service')
 
 const pupilMock = require('../mocks/pupil')
@@ -176,67 +175,6 @@ describe('generate-pins.service', () => {
         const password = school.schoolPin
         const result = generatePinsService.generateSchoolPassword(school)
         expect(result.schoolPin === password).toBeFalsy()
-      })
-    })
-  })
-  describe('getActiveSchool', () => {
-    beforeEach(() => {
-      let school = Object.assign({}, schoolMock)
-      school.pinExpiresAt = moment().startOf('day').add(16, 'hours')
-      sandbox.mock(schoolDataService).expects('findOne').resolves(school)
-      proxyquire('../../services/generate-pins.service', {
-        '../../services/data-access/school.data.service': schoolDataService
-      })
-      describe('if pin is valid', () => {
-        beforeEach(() => {
-          sandbox.useFakeTimers(moment().startOf('day').subtract(1, 'years').valueOf())
-        })
-        it('it should return school object', () => {
-          const result = generatePinsService.getActiveSchool(school.id)
-          expect(result.pinExpiresAt).toBeDefined()
-          expect(result.schoolPin).toBeDefined()
-        })
-      })
-      describe('if pin is invalid', () => {
-        beforeEach(() => {
-          sandbox.useFakeTimers(moment().startOf('day').add(100, 'years').valueOf())
-        })
-        it('it should return null', () => {
-          const result = generatePinsService.getActiveSchool(school.id)
-          expect(result).toBeUndefined()
-        })
-      })
-    })
-  })
-  describe('getPupilsWithActivePins', () => {
-    beforeEach(() => {
-      const pupil1 = Object.assign({}, pupilMock)
-      pupil1.pin = 'f55sg'
-      pupil1.pinExpiresAt = moment().startOf('day').add(16, 'hours')
-      const pupil2 = Object.assign({}, pupilMock)
-      pupil2._id = '595cd5416e5ca13e48ed2520'
-      pupil2.pinExpiresAt = moment().startOf('day').add(16, 'hours')
-      sandbox.mock(pupilDataService).expects('getSortedPupils').resolves([ pupil1, pupil2 ])
-      proxyquire('../../services/generate-pins.service', {
-        '../../services/data-access/pupil.data.service': pupilDataService
-      })
-    })
-    describe('if pins are valid', () => {
-      beforeEach(() => {
-        sandbox.useFakeTimers(moment().startOf('day').subtract(1, 'years').valueOf())
-      })
-      it('it should return a list of active pupils', async () => {
-        const pupils = await generatePinsService.getPupilsWithActivePins(schoolMock._id)
-        expect(pupils.length).toBe(2)
-      })
-    })
-    describe('if pins are invalid', () => {
-      beforeEach(() => {
-        sandbox.useFakeTimers(moment().startOf('day').add(100, 'years').valueOf())
-      })
-      it('it should return a list of active pupils', async () => {
-        const pupils = await generatePinsService.getPupilsWithActivePins(schoolMock._id)
-        expect(pupils.length).toBe(0)
       })
     })
   })
