+++ admin/views/pupil-pin/generate-pins-list.ejs
@@ -17,13 +17,13 @@
 
 <div class="grid-row">
     <div class="column-two-thirds">
-        <form method="post" action="/school/generate-pins" method="post" autocomplete="off">
+        <form method="post" action="/pupil-pin/generate-pins" method="post" autocomplete="off">
             <% if (pupils && pupils.length > 0) { %>
             <table class="spacious sorting" id="generatePins" role="listbox" aria-label="Select pupils.">
                 <thead>
                 <tr>
                     <th scope="col">
-                        <a href="/school/generate-pins-list/lastName/<%= htmlSortDirection.lastName %>">
+                        <a href="/pupil-pin/generate-pins-list/lastName/<%= htmlSortDirection.lastName %>">
                             Pupil <span class="<%= arrowSortDirection.lastName %>"></span>
                         </a>
                     </th>
@@ -78,7 +78,7 @@
 
             <%- partial('partials/_sticky_confirmation_banner', {
                 cancelLabel: 'Cancel',
-                cancelLink: '/school/generate-pins-overview',
+                cancelLink: '/pupil-pin/generate-pins-overview',
                 submitLabel: 'Confirm'
             }) %>
