+++ admin/views/pupil-pin/generate-pins-overview.ejs
@@ -40,7 +40,7 @@
             <li>Generate pupil PINs and school password.</li>
             <li>Distribute pupil PINs and school password to your pupils.</li>
         </ol>
-        <form action="/school/generate-pins-list/lastName/asc" name="generate-pins-list">
+        <form action="/pupil-pin/generate-pins-list/lastName/asc" name="generate-pins-list">
             <div class="form-group">
                 <div class="form-buttons">
                     <input type="submit" class="button" value="Generate PINs"/>
