+++ admin/views/pupil-pin/generated-pins-list.ejs
@@ -59,7 +59,7 @@
     <div class="column-two-thirds">
         <div>
             <div class="icon-file-download pupils-template"></div>
-                <b><a class="template-link" href="/school/print-pins">Download all pupil PINs and school password</a></b>
+                <b><a class="template-link" href="/pupil-pin/print-pins">Download all pupil PINs and school password</a></b>
         </div>
         <br>
         <div>
@@ -73,7 +73,7 @@
 
 <div class="grid-row">
     <div class="column-two-thirds">
-        <form action="/school/generate-pins-list/lastName/asc" name="generate-pins-list">
+        <form action="/pupil-pin/generate-pins-list/lastName/asc" name="generate-pins-list">
             <div class="form-group">
                 <div class="form-buttons">
                     <input type="submit" class="button-secondary" value="Generate more PINs"/>
