+++ admin/views/pupil-pin/pin-print.ejs
@@ -10,7 +10,7 @@
 
             <br/>
 
-            <form action="/school/generate-pins-overview">
+            <form action="/pupil-pin/generate-pins-overview">
                 <div class="form-group">
                     <input id="back" class="button no-print" type="submit" value="Back">
                 </div>
