+++ admin/views/school/school-home.ejs
@@ -35,7 +35,7 @@
             </li>
 
             <li>
-                <h2 class="heading-small"><a href="/school/generate-pins-overview">Generate pupils PINs</a></h2>
+                <h2 class="heading-small"><a href="/pupil-pin/generate-pins-overview">Generate pupils PINs</a></h2>
                 <p>Generate pupil Personal Identification Numbers (PINs) and view school password</p>
             </li>
